UXD-534 refactor panel stories
