+++ packages/Panel/README.md
@@ -41,6 +41,54 @@ npm install @paprika/panel
 | width                 | [string,number]                                                                             | false    | "33%"                         | The width of the open Panel (when slide in from left or right)                                                                                  |
 | zIndex                | number                                                                                      | false    | zValue(7)                     | Control the z-index of the Panel                                                                                                                |
 
+### Panel.Header
+
+| Prop                  | Type                                                    | required | default                   | Description |
+| --------------------- | ------------------------------------------------------- | -------- | ------------------------- | ----------- |
+| children              | node                                                    | true     | -                         |             |
+| getPushContentElement | func                                                    | false    | () => {}                  |             |
+| hasCloseButton        | bool                                                    | false    | true                      |             |
+| isCompact             | bool                                                    | false    | false                     |             |
+| isHeaderSticky        | bool                                                    | false    | false                     |             |
+| kind                  | [ Header.types.kind.DEFAULT, Header.types.kind.PRIMARY] | false    | Header.types.kind.DEFAULT |             |
+| level                 | [ 1, 2, 3, 4, 5, 6]                                     | false    | 2                         |             |
+| onClose               | func                                                    | false    | () => {}                  |             |
+| refHeading            | custom                                                  | false    | null                      |             |
+
+### Panel.Content
+
+| Prop     | Type | required | default | Description                 |
+| -------- | ---- | -------- | ------- | --------------------------- |
+| children | node | false    | null    | Body content of the Content |
+
+### Panel.Footer
+
+| Prop     | Type   | required | default | Description |
+| -------- | ------ | -------- | ------- | ----------- |
+| children | node   | true     | -       |             |
+| height   | number | false    | 72      |             |
+| isSticky | bool   | false    | false   |             |
+
+### Panel.FocusLock
+
+| Prop        | Type | required | default | Description |
+| ----------- | ---- | -------- | ------- | ----------- |
+| returnFocus | -    | false    | true    |             |
+
+### Panel.Overlay
+
+| Prop            | Type   | required | default | Description                                                         |
+| --------------- | ------ | -------- | ------- | ------------------------------------------------------------------- |
+| hasOutsideClick | bool   | false    | true    | Will call an onClose handler when clicked on (outside of the Panel) |
+| onClose         | func   | false    | null    | Callback for click event                                            |
+| zIndex          | number | false    | null    | The z-index of the Panel Overlay                                    |
+
+### Panel.Trigger
+
+| Prop     | Type | required | default | Description |
+| -------- | ---- | -------- | ------- | ----------- |
+| children | node | true     | -       |             |
+
 <!-- autogenerated don't modify -->
 <!-- content -->
 
@@ -49,6 +97,7 @@ npm install @paprika/panel
 You can use any of the following components to compose the Panel:
 
 ```js
+<Panel.Content />
 <Panel.Header />
 <Panel.Footer />
 <Panel.Overlay />
@@ -56,21 +105,24 @@ You can use any of the following components to compose the Panel:
 <Panel.Trigger />
 ```
 
-## Basic examples.
+## Basic examples
 
 ```js
-<Panel isOpen={isOpen} onClose={() => {}}>
-  [...your_content...]
-</Panel>
+import Panel from "@paprika/panel";
+
+const [isOpen, setIsOpen] = React.useState(true);
+
+<Panel isOpen={isOpen} onClose={() => {}} slideFrom={Panel.types.slideFrom.LEFT}>
+  <Panel.Header kind="primary">Header</Panel.Header>
+  <Panel.Content>Your content here</Panel.Content>
+  <Panel.Footer>
+    <button>Accept</button>
+    <button>Cancel</button>
+  </Panel.Footer>
+</Panel>;
 ```
 
-```js
-<Panel isOpen={isOpen} onClose={() => {}}>
-  <Panel.Header>Your header</Panel.Header>
-  [...your_content...]
-  <Panel.Overlay />
-</Panel>
-```
+
 
 ## Warnings
