+++ packages/Panel/package.json
@@ -34,5 +34,15 @@
     "react-dom": "^16.12.0",
     "styled-components": "^4.4.1"
   },
+  "paprikaDocs": {
+    "subComponents": [
+      "Header",
+      "Content",
+      "Footer",
+      "FocusLock",
+      "Overlay",
+      "Trigger"
+    ]
+  },
   "gitHead": "8bff2e8a6f01708ba60d9178f867dba378806370"
 }
