+++ packages/Panel/stories/Panel.examples.stories.js
@@ -1,17 +1,14 @@
 import React from "react";
 import { getStoryName } from "storybook/storyTree";
-import {
-  BasicPanel,
-  FocusLockPanel,
-  FromBottom,
-  GroupPanel,
-  PushContentPanel,
-  StickyFooterPanel,
-  StickyHeaderPanel,
-  ZIndexPanel,
-  FocusHeading,
-  WithSortable,
-} from "./examples/Examples";
+import ExampleStory from "storybook/components/ExampleStory";
+import BasicPanel from "./examples/BasicPanel";
+import FocusHeading from "./examples/FocusHeading";
+import FocusLockPanel from "./examples/FocusLockPanel";
+import FromBottom from "./examples/FromBottom";
+import GroupPanel from "./examples/GroupPanel";
+import PushContentPanel from "./examples/PushContentPanel";
+import StickyPanel from "./examples/StickyPanel";
+
 import Panel from "../src";
 
 const storyName = getStoryName("Panel");
@@ -32,32 +29,58 @@ export default {
   component: Panel,
 };
 
-export const basic = () => <BasicPanel />;
+export const basic = () => (
+  <ExampleStory component="Panel" storyName="Basic Panel" fileName="examples/BasicPanel.js">
+    <BasicPanel />
+  </ExampleStory>
+);
+
 basic.story = sidePanelStoryParameters;
 
-export const focuslock = () => <FocusLockPanel />;
-focuslock.story = sidePanelStoryParameters;
+export const focusHeading = () => (
+  <ExampleStory component="Panel" storyName="Focus Heading Panel" fileName="examples/FocusHeading.js">
+    <FocusHeading />
+  </ExampleStory>
+);
 
-export const group = () => <GroupPanel />;
-group.story = sidePanelStoryParameters;
-
-export const pushContent = () => <PushContentPanel />;
-pushContent.story = sidePanelStoryParameters;
-
-export const stickyFooter = () => <StickyFooterPanel />;
-stickyFooter.story = sidePanelStoryParameters;
-
-export const stickyHeader = () => <StickyHeaderPanel />;
-stickyHeader.story = sidePanelStoryParameters;
-
-export const zIndex = () => <ZIndexPanel />;
-zIndex.story = sidePanelStoryParameters;
-
-export const focusHeading = () => <FocusHeading />;
 focusHeading.story = sidePanelStoryParameters;
 
-export const withSortable = () => <WithSortable />;
-withSortable.story = sidePanelStoryParameters;
+export const focusLock = () => (
+  <ExampleStory component="Panel" storyName="Focus Lock Panel" fileName="examples/FocusLockpanel.js">
+    <FocusLockPanel />
+  </ExampleStory>
+);
 
-export const fromBottom = () => <FromBottom />;
+focusLock.story = sidePanelStoryParameters;
+
+export const fromBottom = () => (
+  <ExampleStory component="Panel" storyName="From Bottom Panel" fileName="examples/FromBottom.js">
+    <FromBottom />
+  </ExampleStory>
+);
+
 fromBottom.story = sidePanelStoryParameters;
+
+export const groupPanel = () => (
+  <ExampleStory component="Panel" storyName="Group Panel" fileName="examples/GroupPanel.js">
+    <GroupPanel />
+  </ExampleStory>
+);
+
+groupPanel.story = sidePanelStoryParameters;
+
+export const pushContentPanel = () => (
+  <ExampleStory component="Panel" storyName="Panel push content" fileName="examples/PushContentPanel.js">
+    <PushContentPanel />
+  </ExampleStory>
+);
+
+pushContentPanel.story = sidePanelStoryParameters;
+
+export const stickyPanel = () => (
+  <ExampleStory component="Panel" storyName="Sticky Panel" fileName="examples/StickyPanel.js">
+    <StickyPanel />
+  </ExampleStory>
+);
+
+stickyPanel.story = sidePanelStoryParameters;
