+++ packages/Panel/stories/tests/PanelDefaultSticky.js
@@ -1,0 +1,22 @@
+import React from "react";
+import Heading from "@paprika/heading";
+import Panel from "../../src";
+import { TextLine, Nav } from "../helpers";
+
+export default function PanelDefaultSticky() {
+  return (
+    <React.Fragment>
+      <Nav />
+      <p>
+        <Panel data-pka-anchor="panel" isOpen offset={{ top: 40 }}>
+          <Panel.Header>
+            <Heading level={2}>Cypress</Heading>
+          </Panel.Header>
+        </Panel>
+      </p>
+      <div style={{ width: "400px", margin: "25px" }}>
+        <TextLine repeat={100} />
+      </div>
+    </React.Fragment>
+  );
+}
